+++ src/app/api/notes/[id]/route.ts
@@ -1,24 +1,31 @@
 import { NextRequest, NextResponse } from 'next/server';
 import dbConnect from '@/lib/mongodb';
-import NoteModel, { INote } from '@/models/NoteModel';
+import NoteModel from '@/models/NoteModel';
 import { Note } from '@/types';
+import { getToken } from 'next-auth/jwt';
 
 interface Params {
   id: string;
 }
 
 export async function PUT(request: NextRequest, { params }: { params: Params }) {
+  const token = await getToken({ req: request, secret: process.env.NEXTAUTH_SECRET });
+  if (!token || !token.id) {
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+  }
+  const userIdAuth = token.id as string;
+
   await dbConnect();
   const { id } = params;
   try {
-    const body: Partial<Omit<Note, 'id'>> = await request.json();
+    const body: Partial<Omit<Note, 'id' | 'userId'>> = await request.json();
     const updateData = {
       ...body,
-      lastEdited: new Date().toISOString(), // Always update lastEdited timestamp
+      lastEdited: new Date().toISOString(),
     };
-    const updatedNote = await NoteModel.findOneAndUpdate({ id: id }, updateData, { new: true, runValidators: true });
+    const updatedNote = await NoteModel.findOneAndUpdate({ id: id, userId: userIdAuth }, updateData, { new: true, runValidators: true });
     if (!updatedNote) {
-      return NextResponse.json({ message: 'Note not found' }, { status: 404 });
+      return NextResponse.json({ message: 'Note not found or you do not have permission to update it.' }, { status: 404 });
     }
     return NextResponse.json(updatedNote, { status: 200 });
   } catch (error) {
@@ -28,12 +35,18 @@ export async function PUT(request: NextRequest, { params }: { params: Params }) 
 }
 
 export async function DELETE(request: NextRequest, { params }: { params: Params }) {
+  const token = await getToken({ req: request, secret: process.env.NEXTAUTH_SECRET });
+  if (!token || !token.id) {
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+  }
+  const userIdAuth = token.id as string;
+
   await dbConnect();
   const { id } = params;
   try {
-    const deletedNote = await NoteModel.findOneAndDelete({ id: id });
+    const deletedNote = await NoteModel.findOneAndDelete({ id: id, userId: userIdAuth });
     if (!deletedNote) {
-      return NextResponse.json({ message: 'Note not found' }, { status: 404 });
+      return NextResponse.json({ message: 'Note not found or you do not have permission to delete it.' }, { status: 404 });
     }
     return NextResponse.json({ message: 'Note deleted successfully' }, { status: 200 });
   } catch (error) {
@@ -41,3 +54,4 @@ export async function DELETE(request: NextRequest, { params }: { params: Params 
     return NextResponse.json({ message: `Failed to delete note ${id}`, error: (error as Error).message }, { status: 500 });
   }
 }
+
